Started to move access_control method out of namespace iris_engine into business layer
