+++ source/app/blueprints/rest/manage/manage_assets_routes.py
@@ -21,7 +21,7 @@ from flask import request
 from werkzeug import Response
 
 from app.datamgmt.manage.manage_assets_db import get_filtered_assets
-from app.iris_engine.access_control.utils import ac_fast_check_current_user_has_case_access
+from app.business.access_controls import ac_fast_check_current_user_has_case_access
 from app.models.authorization import CaseAccessLevel
 from app.schema.marshables import CaseAssetsSchema
 from app.blueprints.access_controls import ac_api_requires
