+++ source/app/blueprints/rest/v2/tasks.py
@@ -30,7 +30,7 @@ from app.business.errors import ObjectNotFoundError
 from app.business.errors import BusinessProcessingError
 from app.models.authorization import CaseAccessLevel
 from app.schema.marshables import CaseTaskSchema
-from app.iris_engine.access_control.utils import ac_fast_check_current_user_has_case_access
+from app.business.access_controls import ac_fast_check_current_user_has_case_access
 from app.blueprints.rest.v2.tasks_routes.comments import tasks_comments_blueprint
